+++ lib/ruby_challenge/models/segments/base.rb
@@ -3,12 +3,12 @@
 module RubyChallenge
   module Segments
     class Base
-      attr_reader :origin, :departure, :arrival, :errors
+      attr_reader :origin, :from_time, :to_time, :errors
 
-      def initialize(origin:, departure:, arrival:)
+      def initialize(origin:, from_time:, to_time:)
         @origin = origin
-        @departure = departure
-        @arrival = arrival
+        @from_time = from_time
+        @to_time = to_time
         @errors = []
       end
 
@@ -29,11 +29,11 @@ module RubyChallenge
       private
 
       def check_errors_for_datetimes
-        @errors << "departure is not valid" if @departure.nil? || !@departure.is_a?(Time)
-        @errors << "arrival is not valid" if @arrival.nil? || !@arrival.is_a?(Time)
-        return if !@arrival.is_a?(Time) || !@departure.is_a?(Time)
+        @errors << "from_time is not valid" if @from_time.nil? || !@from_time.is_a?(Time)
+        @errors << "to_time is not valid" if @to_time.nil? || !@to_time.is_a?(Time)
+        return if !@to_time.is_a?(Time) || !@from_time.is_a?(Time)
 
-        @errors << "arrival cannot be lesser than departure" if @arrival < @departure
+        @errors << "to_time cannot be lesser than from_time" if @to_time < @from_time
       end
     end
   end