+++ lib/ruby_challenge/models/segments/flight.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RubyChallenge
+  module Segments
+    class Flight < Transport; end
+  end
+end