+++ lib/ruby_challenge/models/trip.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module RubyChallenge
+  class Trip
+    def initialize
+      @potato = nil
+    end
+  end
+end