+++ spec/models/segments/hotel_spec.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
-RSpec.describe RubyChallenge::Segments::Base do
+RSpec.describe RubyChallenge::Segments::Hotel do
+  let(:custom_from_time) { Time.new(2024, 2, 1, 12, 0, 0) }
+  let(:three_days) { 3 * 24 * 3600 }
   let(:origin) { "BCN" }
-  let(:from_time) { Time.now }
-  let(:to_time) { Time.now + (3 * 3600) }
+  let(:from_time) { custom_from_time }
+  let(:to_time) { custom_from_time + three_days }
 
   subject { described_class.new(origin:, from_time:, to_time:) }
 
@@ -20,7 +22,7 @@ RSpec.describe RubyChallenge::Segments::Base do
     end
 
     it "run error when call to_s" do
-      expect { subject.to_s }.to raise_error(NotImplementedError)
+      expect(subject.to_s).to eq("Hotel at BCN on 2024-02-01 to 2024-02-04")
     end
   end