+++ spec/models/segments/train_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe RubyChallenge::Segments::Train do
+  include_context "transport segment model"
+end